Separate classic and taskflow example dags for setup/teardown
