+++ airflow/example_dags/example_setup_teardown.py
@@ -20,7 +20,6 @@ from __future__ import annotations
 
 import pendulum
 
-from airflow.decorators import setup, task, task_group, teardown
 from airflow.models.dag import DAG
 from airflow.operators.bash import BashOperator
 from airflow.utils.task_group import TaskGroup
@@ -36,46 +35,10 @@ with DAG(
     BashOperator.as_teardown(task_id="root_teardown", bash_command="echo 'Goodbye from root_teardown'")
 
     with TaskGroup("section_1") as section_1:
+        BashOperator.as_setup(task_id="taskgroup_setup", bash_command="echo 'Hello from taskgroup_setup'")
+        BashOperator(task_id="normal", bash_command="echo 'I am just a normal task'")
+        BashOperator.as_setup(
+            task_id="taskgroup_teardown", bash_command="echo 'Hello from taskgroup_teardown'"
+        )
 
-        @setup
-        @task
-        def my_setup():
-            print("I set up")
-
-        @task
-        def hello():
-            print("I say hello")
-
-        @teardown
-        @task
-        def my_teardown():
-            print("I tear down")
-
-        my_setup()
-        hello()
-        my_teardown()
-
-    with TaskGroup("section_2") as section_2:
-
-        @setup
-        @task_group
-        def my_setup_taskgroup():
-            @task
-            def first_setup():
-                print("I set some stuff up")
-
-            @task
-            def second_setup():
-                print("I set some other stuff up")
-
-            first_setup()
-            second_setup()
-
-        @task
-        def hello():
-            print("I say hello")
-
-        my_setup_taskgroup()
-        hello()
-
-    normal >> section_1 >> section_2
+    normal >> section_1
